+++ packages/hwa/src/Wallet/index.ts
@@ -8,20 +8,20 @@ import {
   isValidXAddress,
   xAddressToClassicAddress,
   encodeSeed,
-} from 'ripple-address-codec'
+} from 'hwa-ripple-address-codec'
 import {
   decode,
   encodeForSigning,
   encodeForMultisigning,
   encode,
-} from 'ripple-binary-codec'
+} from 'hwa-ripple-binary-codec'
 import {
   deriveAddress,
   deriveKeypair,
   generateSeed,
   verify,
   sign,
-} from 'ripple-keypairs'
+} from 'hwa-ripple-keypairs'
 
 import ECDSA from '../ECDSA'
 import { ValidationError, XrplError } from '../errors'