+++ packages/hwa/webpack.config.js
@@ -60,7 +60,7 @@ function getDefaultConfiguration() {
       },
       extensions: ['.js', '.json'],
       // We don't want to webpack any of the local dependencies:
-      // ripple-address-codec, ripple-binary-codec, ripple-keypairs, which are
+      // hwa-ripple-address-codec, hwa-ripple-binary-codec, hwa-ripple-keypairs, which are
       // symlinked together via lerna
       symlinks: false,
       fallback: {