+++ packages/ripple-address-codec/HISTORY.md
@@ -1,4 +1,4 @@
-# ripple-address-codec
+# hwa-ripple-address-codec
 
 ## Unreleased
 ### Changed
@@ -13,8 +13,8 @@
 - Initial pass at linting this codebase with new rules
 
 ## 4.2.0 (2021-11-15)
-- Converts ripple-address-codec into a monorepo with ripple-binary-codec,
-  ripple-keypairs, and xrpl. Changes to build tooling but no new features or
+- Converts hwa-ripple-address-codec into a monorepo with hwa-ripple-binary-codec,
+  hwa-ripple-keypairs, and xrpl. Changes to build tooling but no new features or
   bug fixes
 
 ## 4.1.3 (2021-05-10)